+++ common/src/main/java/io/github/stainlessstasis/core/CobblemonSpawnAlerts.java
@@ -13,14 +13,16 @@ import com.cobblemon.mod.common.pokemon.Nature;
 import com.cobblemon.mod.common.pokemon.Pokemon;
 import com.cobblemon.mod.common.pokemon.stat.CobblemonStatProvider;
 import io.github.stainlessstasis.alert.DespawnReason;
-import io.github.stainlessstasis.config.CommonConfigManager;
-import io.github.stainlessstasis.config.ServerConfig;
+import io.github.stainlessstasis.config.manager.CommonConfigManager;
+import io.github.stainlessstasis.config.common.ServerConfig;
+import io.github.stainlessstasis.config.manager.VersionMatcher;
 import io.github.stainlessstasis.network.*;
 import io.github.stainlessstasis.platform.Services;
 import io.github.stainlessstasis.util.*;
 import kotlin.Unit;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
+import org.jetbrains.annotations.ApiStatus;
 import org.joml.Vector3f;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -81,6 +83,10 @@ public class CobblemonSpawnAlerts {
 
             return Unit.INSTANCE;
         });
+    }
+
+    public static String getLastKnownModVersion() {
+        return VersionMatcher.getLastKnownModVersion();
     }
 
     public static PokemonDataPacket createPokemonData(PokemonEntity pokemonEntity, RarityUtil.Bucket bucket) {